# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventManagement swagger application",
        "description": "Api d'une Application de  gestion des evenements ",
        "contact": {
            "email": "taneyemc@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/owner/addOwner": {
            "post": {
                "tags": [
                    "owner"
                ],
                "summary": "Add owner",
                "description": "Create a new owner",
                "operationId": "d9f7f4799ecd9bcd2699a59f17ab65bb",
                "requestBody": {
                    "description": "Owner details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "firstname": {
                                        "description": "Owner's firstname",
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "description": "Owner's lastname",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Owner's email",
                                        "type": "string"
                                    },
                                    "gender": {
                                        "description": "Owner's gender",
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "description": "Owner's adresse",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Owner's phone",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Owner's description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Owner added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Owner added successfully"
                                        },
                                        "owner": {
                                            "$ref": "#/components/schemas/Owner"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "firstname": [
                                                    "The firstname field is required."
                                                ],
                                                "email": [
                                                    "The email field is required.",
                                                    "The email field must be in the form of an email address.",
                                                    "there is already a user with this email address, please try another"
                                                ],
                                                "phone": [
                                                    "The phone field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/owner/updateOwner/{owner}": {
            "put": {
                "tags": [
                    "owner"
                ],
                "summary": "Update an Owner",
                "description": "Update an existing owner by ID",
                "operationId": "a9f86ac4dbafff56714eef21770b3e6a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the owner",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Owner details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "firstname": {
                                        "description": "Owner's firstname",
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "description": "Owner's lastname",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Owner's email",
                                        "type": "string"
                                    },
                                    "gender": {
                                        "description": "Owner's gender",
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "description": "Owner's adresse",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Owner's phone",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Owner's description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Owner updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Owner updated successfully"
                                        },
                                        "owner": {
                                            "$ref": "#/components/schemas/Owner"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Owner not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "firstname": [
                                                    "The firstname field is required."
                                                ],
                                                "email": [
                                                    "The email field is required.",
                                                    "The email field must be in the form of an email address."
                                                ],
                                                "phone": [
                                                    "The phone field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/owner/getAllOwner": {
            "get": {
                "tags": [
                    "owner"
                ],
                "summary": "Get all Owners",
                "description": "Retrieve a list of all owners",
                "operationId": "30f42cc290b7e2a78706b1b0af868d82",
                "responses": {
                    "200": {
                        "description": "List of all owners",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Owner"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No owners found"
                    }
                }
            }
        },
        "/api/owner/showOwner/{owner}": {
            "get": {
                "tags": [
                    "owner"
                ],
                "summary": "Show Owner detail by ID",
                "description": "Retrieve details of a specific owner by ID",
                "operationId": "66ebaaa14e416a6fdd0f9d3070adb329",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the owner",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Owner details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Owner"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Owner not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Owner": {
                "required": [
                    "firstname",
                    "email",
                    "phone"
                ],
                "properties": {
                    "firstname": {
                        "description": "Owner's firstname",
                        "type": "string"
                    },
                    "lastname": {
                        "description": "Owner's lastname",
                        "type": "string"
                    },
                    "email": {
                        "description": "Owner's email",
                        "type": "string"
                    },
                    "gender": {
                        "description": "Owner's gender",
                        "type": "string"
                    },
                    "adresse": {
                        "description": "Owner's adresse",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Owner's phone",
                        "type": "string"
                    },
                    "description": {
                        "description": "Owner's description",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}